Seguindo com a Implementacao do Identity

+++ GerenciamentoPlantao/Services/AcionamentoService.cs
@@ -1,6 +1,7 @@
 ﻿using GerenciamentoPlantao.Data;
 using GerenciamentoPlantao.Models;
 using GerenciamentoPlantao.Models.ViewModels;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.EntityFrameworkCore;
 
 namespace GerenciamentoPlantao.Services
@@ -8,10 +9,12 @@ namespace GerenciamentoPlantao.Services
     public class AcionamentoService
     {
         private readonly GerenciamentoPlantaoContext _context;
+        private readonly IUsuarioLogadoService _usuarioLogado;
 
-        public AcionamentoService(GerenciamentoPlantaoContext context)
+        public AcionamentoService(GerenciamentoPlantaoContext context, IUsuarioLogadoService usuarioLogado)
         {
             _context = context;
+            _usuarioLogado = usuarioLogado;
         }
 
         public async Task<List<Acionamento>> FindAllAsync()
@@ -41,11 +44,18 @@ namespace GerenciamentoPlantao.Services
 
         public async Task InsertAsync(AcionamentoFormViewModel vm)
         {
+            var usuario = await _usuarioLogado.ObterUsuarioLogadoAsync();
+            
+            if (usuario == null)
+            {
+                throw new Exception("Usuário não encontrado.");
+            }   
+
             var acionamento = new Acionamento
             {
                 DataRegistro = DateTime.Now,
                 DataAcionamento = vm.DataAcionamento,
-                UsuarioId = vm.UsuarioId,
+                UsuarioId = usuario.Id,
                 CanalId = vm.CanalId,
                 Acionador = vm.Acionador,
                 NrAtendimento = vm.NrAtendimento,
@@ -63,15 +73,25 @@ namespace GerenciamentoPlantao.Services
 
         public async Task UpdateAsync(EditarAcionamentoViewModel vm)
         {
+            var usuario = await _usuarioLogado.ObterUsuarioLogadoAsync();
+            if (usuario == null)
+            {
+                throw new Exception("Usuário não autenticado.");
+            }
+
             var acionamento = await _context.Acionamentos.FindAsync(vm.Id);
             if (acionamento == null)
             {
                 throw new Exception("Acionamento não encontrado.");
             }
+
+            if (acionamento.UsuarioId != usuario.Id)
+            {
+                throw new Exception("Você não tem permissão para editar este acionamento.");
+            }
             acionamento.AtualizarAcionamento(
                 vm.DataAcionamento,
-                vm.CanalId,
-                vm.UsuarioId,   
+                vm.CanalId,  
                 vm.EstabelecimentoId,
                 vm.SetorId,
                 vm.CategoriaId,
@@ -81,15 +101,24 @@ namespace GerenciamentoPlantao.Services
                 vm.Acionador,
                 vm.NrAtendimento
             );
+
             _context.Acionamentos.Update(acionamento);
             await _context.SaveChangesAsync();
         }
 
         public async Task DeleteAsync(int id) {
+            
+            var usuario = await _usuarioLogado.ObterUsuarioLogadoAsync();
             var acionamento = await _context.Acionamentos.FindAsync(id);
+
             if (acionamento == null)
             {
                 throw new Exception("Acionamento não encontrado.");
+            }
+
+            if (acionamento.UsuarioId != usuario.Id)
+            {
+                throw new Exception("Você não tem permissão para excluir este acionamento.");
             }
             _context.Acionamentos.Remove(acionamento);
             await _context.SaveChangesAsync();
